Move all validation functions into its own module
This will make further refactoring easier and will generally be cleaner
from an organizational standpoint.
