+++ gridpath/project/capacity/capacity_types/stor_new_bin.py
@@ -19,21 +19,18 @@ incurred by binary new-build storage.
 
 from __future__ import print_function
 
-from builtins import next
 import csv
 import os.path
 import pandas as pd
-import numpy as np
 from pyomo.environ import Set, Param, Var, NonNegativeReals, \
     Constraint, value, Binary
 
-from db.common_functions import spin_on_database_lock
-from gridpath.auxiliary.auxiliary import check_column_sign_positive, \
-    get_expected_dtypes, check_dtypes, write_validation_to_database, \
-    setup_results_import
 from gridpath.auxiliary.dynamic_components import \
     capacity_type_operational_period_sets, \
     storage_only_capacity_type_operational_period_sets
+from gridpath.auxiliary.validations import write_validation_to_database, \
+    get_expected_dtypes, check_dtypes, check_column_sign_positive, \
+    validate_projects, validate_costs
 from gridpath.project.capacity.capacity_types.common_methods import \
     operational_periods_by_project_vintage, project_operational_periods, \
     project_vintages_operational_in_period, update_capacity_results_table
@@ -746,39 +743,3 @@ def validate_module_specific_inputs(subscenarios, subproblem, stage, conn):
         errors=validate_costs(cost_df, prj_vintages)
     )
 
-
-def validate_projects(list1, list2):
-    """
-    Check for projects in list 1 that aren't in list 2
-
-    Note: Function will ignore duplicates in list
-    :param list1:
-    :param list2:
-    :return: list of error messages for each missing project
-    """
-    results = []
-    missing_projects = np.setdiff1d(list1, list2)
-    for prj in missing_projects:
-        results.append(
-            "Missing build size inputs for project '{}'".format(prj)
-        )
-
-    return results
-
-
-def validate_costs(cost_df, prj_vintages):
-    """
-    Check that cost inputs exist for the relevant projects and vintages
-    :param cost_df:
-    :param prj_vintages: list with relevant projects and vintages (tuple)
-    :return: list of error messages for each missing project, period combination
-    """
-    results = []
-    for prj, vintage in prj_vintages:
-        if not ((cost_df.project == prj) & (cost_df.period == vintage)).any():
-            results.append(
-                "Missing cost inputs for project '{}', vintage '{}'"
-                .format(prj, str(vintage))
-            )
-
-    return results
