+++ gridpath/project/operations/__init__.py
@@ -16,9 +16,10 @@ import pandas as pd
 import os.path
 from pyomo.environ import Set, Param, PositiveReals, Reals, NonNegativeReals
 
-from gridpath.auxiliary.auxiliary import check_dtypes, \
-    get_expected_dtypes, check_column_sign_positive, \
-    write_validation_to_database
+from gridpath.auxiliary.validations import write_validation_to_database, \
+    check_dtypes, get_expected_dtypes, check_column_sign_positive, \
+    validate_fuel_vs_heat_rates, validate_heat_rate_curves, \
+    validate_vom_curves, validate_min_stable_level
 from gridpath.project.common_functions import append_to_input_file
 
 
@@ -822,198 +823,3 @@ def validate_inputs(subscenarios, subproblem, stage, conn):
 
     # TODO: check that if there is a "0" for the period for a given
     #  project there are zeroes everywhere for that project.
-
-
-def validate_min_stable_level(df):
-    """
-    Check 0 < min stable fraction <= 1
-    :param df:
-    :return:
-    """
-    results = []
-
-    invalids = ((df["min_stable_level_fraction"] <= 0) |
-                (df["min_stable_level_fraction"] > 1))
-    if invalids.any():
-        bad_projects = df["project"][invalids].values
-        print_bad_projects = ", ".join(bad_projects)
-        results.append(
-            "Project(s) '{}': expected 0 < min_stable_level <= 1"
-            .format(print_bad_projects)
-        )
-
-    return results
-
-
-def validate_fuel_vs_heat_rates(hr_df):
-    """
-    Make sure projects with fuel have a heat rate scenario specified.
-    Conversely, if no fuel is specified, make sure there is no heat rate
-    scenario specified.
-    :param hr_df:
-    :return:
-    """
-    results = []
-
-    hr_curve_mask = pd.notna(hr_df["heat_rate_curves_scenario_id"])
-    fuel_mask = pd.notna(hr_df["fuel"])
-
-    invalids = fuel_mask & ~hr_curve_mask
-    if invalids.any():
-        bad_projects = hr_df["project"][invalids]
-        print_bad_projects = ", ".join(bad_projects)
-        results.append(
-            "Project(s) '{}': Missing heat_rate_curves_scenario_id"
-            .format(print_bad_projects)
-        )
-
-    invalids = ~fuel_mask & hr_curve_mask
-    if invalids.any():
-        bad_projects = pd.unique(hr_df["project"][invalids])
-        print_bad_projects = ", ".join(bad_projects)
-        results.append(
-             "Project(s) '{}': No fuel specified so no heat rate expected"
-             .format(print_bad_projects)
-        )
-
-    return results
-
-
-def validate_heat_rate_curves(hr_df):
-    """
-    1. Check that specified heat rate scenarios actually have inputs in the
-       heat rate curves table
-    2. Check that specified heat rate curves inputs are valid:
-        - strictly increasing load points
-        - increasing total fuel burn
-        - convex fuel burn curve
-    :param hr_df:
-    :return:
-    """
-    results = []
-
-    fuel_mask = pd.notna(hr_df["fuel"])
-    hr_curve_mask = pd.notna(hr_df["heat_rate_curves_scenario_id"])
-    load_point_mask = pd.notna(hr_df["load_point_fraction"])
-
-    # Check for missing inputs in heat rates curves table
-    invalids = hr_curve_mask & ~load_point_mask
-    if invalids.any():
-        bad_projects = hr_df["project"][invalids]
-        print_bad_projects = ", ".join(bad_projects)
-        results.append(
-            "Project(s) '{}': Expected at least one load point"
-            .format(print_bad_projects)
-        )
-
-    # Check that each project has convex heat rates etc.
-    relevant_mask = fuel_mask & load_point_mask
-    hr_df = hr_df[relevant_mask]
-    for project in hr_df["project"].unique():
-        for period in hr_df[hr_df["project"] == project]["period"].unique():
-            # read in the power setpoints and average heat rates
-            hr_slice = hr_df[(hr_df["project"] == project)
-                             & (hr_df["period"] == period)]
-            hr_slice = hr_slice.sort_values(by=["load_point_fraction"])
-            load_points = hr_slice["load_point_fraction"].values
-            heat_rates = hr_slice["average_heat_rate_mmbtu_per_mwh"].values
-
-            if len(load_points) > 1:
-                incr_loads = np.diff(load_points)
-
-                if np.any(incr_loads == 0):
-                    # note: primary key should already prohibit this
-                    results.append(
-                        "Project(s) '{}': load points can not be identical"
-                        .format(project)
-                    )
-
-                else:
-                    fuel_burn = load_points * heat_rates
-                    incr_fuel_burn = np.diff(fuel_burn)
-                    slopes = incr_fuel_burn / incr_loads
-
-                    if np.any(incr_fuel_burn <= 0):
-                        results.append(
-                            "Project(s) '{}': Total fuel burn should increase "
-                            "with increasing load"
-                            .format(project)
-                        )
-                    if np.any(np.diff(slopes) <= 0):
-                        results.append(
-                            "Project(s) '{}': Fuel burn should be convex, "
-                            "i.e. marginal heat rate should increase with "
-                            "increading load"
-                            .format(project)
-                        )
-
-    return results
-
-
-def validate_vom_curves(vom_df):
-    """
-    1. Check that specified variable O&M scenarios actually have inputs in the
-       variable O&M curves table
-    2. Check that specified variable O&M curves inputs are valid:
-        - strictly increasing load points
-        - increasing total variable O&M cost
-        - convex variable O&M curve
-    :param vom_df:
-    :return:
-    """
-    results = []
-
-    vom_curve_mask = pd.notna(vom_df["variable_om_curves_scenario_id"])
-    load_point_mask = pd.notna(vom_df["load_point_fraction"])
-
-    # Check for missing inputs in heat rates curves table
-    invalids = vom_curve_mask & ~load_point_mask
-    if invalids.any():
-        bad_projects = vom_df["project"][invalids]
-        print_bad_projects = ", ".join(bad_projects)
-        results.append(
-            "Project(s) '{}': Expected at least one load point"
-            .format(print_bad_projects)
-        )
-
-    # Check that each project has convex variable O&M rates etc.
-    vom_df = vom_df[load_point_mask]
-    for project in vom_df["project"].unique():
-        for period in vom_df[vom_df["project"] == project]["period"].unique():
-            # read in the power setpoints and average variable O&M
-            vom_slice = vom_df[(vom_df["project"] == project)
-                               & (vom_df["period"] == period)]
-            vom_slice = vom_slice.sort_values(by=["load_point_fraction"])
-            load_points = vom_slice["load_point_fraction"].values
-            vom = vom_slice["average_variable_om_cost_per_mwh"].values
-
-            if len(load_points) > 1:
-                incr_loads = np.diff(load_points)
-
-                if np.any(incr_loads == 0):
-                    # note: primary key should already prohibit this
-                    results.append(
-                        "Project(s) '{}': load points can not be identical"
-                        .format(project)
-                    )
-
-                else:
-                    vom_cost = load_points * vom
-                    incr_vom_cost = np.diff(vom_cost)
-                    slopes = incr_vom_cost / incr_loads
-
-                    if np.any(incr_vom_cost <= 0):
-                        results.append(
-                            "Project(s) '{}': Total variable O&M cost should "
-                            "increase with increasing load"
-                            .format(project)
-                        )
-                    if np.any(np.diff(slopes) <= 0):
-                        results.append(
-                            "Project(s) '{}': Variable O&M cost should be "
-                            "convex, i.e. variable O&M rate should increase "
-                            "with increasing load"
-                            .format(project)
-                        )
-
-    return results
