+++ gridpath/project/operations/operational_types/gen_var_must_take.py
@@ -11,9 +11,9 @@ of this operational type cannot provide operational reserves .
 from pyomo.environ import Param, Set, NonNegativeReals, Constraint
 import warnings
 
-from gridpath.auxiliary.auxiliary import generator_subset_init, \
-    write_validation_to_database, get_projects_by_reserve, \
-    check_projects_for_reserves
+from gridpath.auxiliary.auxiliary import generator_subset_init
+from gridpath.auxiliary.validations import write_validation_to_database, \
+    get_projects_by_reserve, check_projects_for_reserves
 from gridpath.auxiliary.dynamic_components import headroom_variables, \
     footroom_variables
 from gridpath.project.common_functions import \
@@ -370,7 +370,6 @@ def validate_module_specific_inputs(subscenarios, subproblem, stage, conn):
                                  "gen_var_must_take")
 
     # Other module specific validations
-    validation_results = []
 
     # TODO: validate timepoints: make sure timepoints specified are consistent
     #   with the temporal timepoints (more is okay, less is not)
