+++ gridpath/validate_inputs.py
@@ -16,8 +16,8 @@ import sys
 from argparse import ArgumentParser
 
 from db.common_functions import connect_to_database, spin_on_database_lock
-from gridpath.auxiliary.auxiliary import get_scenario_id_and_name, \
-    write_validation_to_database
+from gridpath.auxiliary.auxiliary import get_scenario_id_and_name
+from gridpath.auxiliary.validations import write_validation_to_database
 from gridpath.common_functions import get_db_parser
 from gridpath.auxiliary.module_list import determine_modules, load_modules
 from gridpath.auxiliary.scenario_chars import OptionalFeatures, SubScenarios, \
