+++ tests/project/operations/test_init.py
@@ -524,92 +524,6 @@ class TestOperationsInit(unittest.TestCase):
             self.assertListEqual(expected_intercepts.tolist(),
                                  actual_intercepts.tolist())
 
-    def test_project_validations(self):
-        cols = ["project", "min_stable_level_fraction"]
-        test_cases = {
-            # Make sure correct inputs don't throw error
-            1: {"df": pd.DataFrame(
-                    columns=cols,
-                    data=[["gas_ct", 0.5]
-                          ]),
-                "min_stable_level_error": [],
-                },
-            # Make sure invalid min_stable_level is flagged
-            2: {"df": pd.DataFrame(
-                columns=cols,
-                data=[["gas_ct1", 1.5],
-                      ["gas_ct2", 0]
-                      ]),
-                "min_stable_level_error": ["Project(s) 'gas_ct1, gas_ct2': expected 0 < min_stable_level <= 1"],
-                }
-        }
-
-        for test_case in test_cases.keys():
-            expected_list = test_cases[test_case]["min_stable_level_error"]
-            actual_list = MODULE_BEING_TESTED.validate_min_stable_level(
-                df=test_cases[test_case]["df"]
-            )
-            self.assertListEqual(expected_list, actual_list)
-
-    def test_heat_rate_validations(self):
-        hr_columns = ["project", "fuel", "heat_rate_curves_scenario_id",
-                      "period",
-                      "load_point_fraction", "average_heat_rate_mmbtu_per_mwh"]
-        test_cases = {
-            # Make sure correct inputs don't throw error
-            1: {"hr_df": pd.DataFrame(
-                    columns=hr_columns,
-                    data=[["gas_ct", "gas", 1, 2020, 10, 10.5],
-                          ["gas_ct", "gas", 1, 2020, 20, 9],
-                          ["coal_plant", "coal", 1, 2020, 100, 10]
-                          ]),
-                "fuel_vs_hr_error": [],
-                "hr_curves_error": []
-                },
-            # Check fuel vs heat rate curve errors
-            2: {"hr_df": pd.DataFrame(
-                columns=hr_columns,
-                data=[["gas_ct", "gas", None, None, None, None],
-                      ["coal_plant", None, 1, 2020, 100, 10]
-                      ]),
-                "fuel_vs_hr_error": ["Project(s) 'gas_ct': Missing heat_rate_curves_scenario_id",
-                                     "Project(s) 'coal_plant': No fuel specified so no heat rate expected"],
-                "hr_curves_error": []
-                },
-            # Check heat rate curves validations
-            3: {"hr_df": pd.DataFrame(
-                columns=hr_columns,
-                data=[["gas_ct1", "gas", 1, None, None, None],
-                      ["gas_ct2", "gas", 1, 2020, 10, 11],
-                      ["gas_ct2", "gas", 1, 2020, 10, 12],
-                      ["gas_ct3", "gas", 1, 2020, 10, 11],
-                      ["gas_ct3", "gas", 1, 2020, 20, 5],
-                      ["gas_ct4", "gas", 1, 2020, 10, 11],
-                      ["gas_ct4", "gas", 1, 2020, 20, 10],
-                      ["gas_ct4", "gas", 1, 2020, 30, 9]
-                      ]),
-                "fuel_vs_hr_error": [],
-                "hr_curves_error": ["Project(s) 'gas_ct1': Expected at least one load point",
-                                    "Project(s) 'gas_ct2': load points can not be identical",
-                                    "Project(s) 'gas_ct3': Total fuel burn should increase with increasing load",
-                                    "Project(s) 'gas_ct4': Fuel burn should be convex, i.e. marginal heat rate should increase with increading load"]
-                },
-
-        }
-
-        for test_case in test_cases.keys():
-            expected_list = test_cases[test_case]["fuel_vs_hr_error"]
-            actual_list = MODULE_BEING_TESTED.validate_fuel_vs_heat_rates(
-                hr_df=test_cases[test_case]["hr_df"]
-            )
-            self.assertListEqual(expected_list, actual_list)
-
-            expected_list = test_cases[test_case]["hr_curves_error"]
-            actual_list = MODULE_BEING_TESTED.validate_heat_rate_curves(
-                hr_df=test_cases[test_case]["hr_df"]
-            )
-            self.assertListEqual(expected_list, actual_list)
-
 
 if __name__ == "__main__":
     unittest.main()
